Added a stub unit test to the source template to avoid jest failing on fresh project.
When initializing an ASIDE project, the CLI would not generate a test folder. That leads
to any command including the `test` command to fail as Jest complains about the test folder
not being there. Modified the template folder to include a stub test that gets generated
alongside the stub source.